+++ internal/application/usecase/checkbalance.go
@@ -8,8 +8,8 @@ import (
 )
 
 type Balance struct {
-	Current   int `json:"current"`
-	Withdrawn int `json:"withdrawn"`
+	Current   float32 `json:"current"`
+	Withdrawn float32 `json:"withdrawn"`
 }
 
 func CheckBalance(
