+++ migrations/20220728194635_create_withdraws_table.up.sql
@@ -3,6 +3,6 @@ CREATE TABLE IF NOT EXISTS withdraws
     id UUID NOT NULL PRIMARY KEY,
     order_number VARCHAR(255) NOT NULL CONSTRAINT withdraws_orders_fk REFERENCES orders(number),
     user_id UUID NOT NULL CONSTRAINT withdraws_users_fk REFERENCES users(id),
-    sum INTEGER NOT NULL,
+    sum NUMERIC(6,2) NOT NULL,
     processed_at TIMESTAMP  NOT NULL
 );
